# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PmsAssistantQuestionInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  question: string;
}

export interface PmsAssistantStatus {
  ready: boolean;
  reason: string;
}

export type PmsAssistantAnswerStatus = typeof PmsAssistantAnswerStatus[keyof typeof PmsAssistantAnswerStatus];


export const PmsAssistantAnswerStatus = {
  answered: 'answered',
  unavailable: 'unavailable',
} as const;

export type PmsAssistantCitationKind = typeof PmsAssistantCitationKind[keyof typeof PmsAssistantCitationKind];


export const PmsAssistantCitationKind = {
  client: 'client',
  project: 'project',
  task: 'task',
} as const;

export interface PmsAssistantCitation {
  kind: PmsAssistantCitationKind;
  id: string;
  title: string;
  excerpt: string;
}

export type PmsAssistantClaimBasis = typeof PmsAssistantClaimBasis[keyof typeof PmsAssistantClaimBasis];


export const PmsAssistantClaimBasis = {
  recorded: 'recorded',
  inference: 'inference',
} as const;

export interface PmsAssistantClaim {
  text: string;
  basis: PmsAssistantClaimBasis;
  /** @minItems 1 */
  citationIds: string[];
}

export interface PmsAssistantAnswer {
  status: PmsAssistantAnswerStatus;
  answer: string;
  citations: PmsAssistantCitation[];
  claims: PmsAssistantClaim[];
}

export interface ApiError {
  error: string;
}