# lib/api-zod/src/generated/types/pmsAssistantAnswer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PmsAssistantAnswerStatus } from './pmsAssistantAnswerStatus';
import type { PmsAssistantCitation } from './pmsAssistantCitation';
import type { PmsAssistantClaim } from './pmsAssistantClaim';

export interface PmsAssistantAnswer {
  status: PmsAssistantAnswerStatus;
  answer: string;
  citations: PmsAssistantCitation[];
  claims: PmsAssistantClaim[];
}