# lib/api-zod/src/generated/types/pmsAssistantCitation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PmsAssistantCitationKind } from './pmsAssistantCitationKind';

export interface PmsAssistantCitation {
  kind: PmsAssistantCitationKind;
  id: string;
  title: string;
  excerpt: string;
}