# lib/api-zod/src/generated/types/pmsAssistantClaim.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PmsAssistantClaimBasis } from './pmsAssistantClaimBasis';

export interface PmsAssistantClaim {
  text: string;
  basis: PmsAssistantClaimBasis;
  /** @minItems 1 */
  citationIds: string[];
}